Added resources to the add software page.
Tests need to be updated due to the way the software controller has been
adapted to deal with the way new software pages are submitted (namely, the
controller now returns either html or json, but the tests only test the html
responses).

Added helper javascript for verticals (currently for software, but should be
usable by other verticals).

Removed turbolinks because it was interferring with ajax.