Added an interface for advanced search.
The same interface is used for the header as on the search page. The search
bar in the header is no longer exposed on the search page. On a SERP, clicking
on a vertical tab will send the original search request (as spelled out in
the URL) to the clicked vertical.